tetris.cvm: Make rotation work better.
I didn't follow any specific guides on how rotation is supposed to work
in tetris other than doing some research and finding out it's not even
particularly consistent within mainline tetris.  I'm OK with it working
the way it does as this is just a demonstration game anyway so if it's a
bit quirky or inaccurate, that's OK.  It's just much more playable now.