crustyvm.c: Make floats more useful.
Allow defining size of array to allow for large arrays of floats without
having to define every member.  Defaults to 0.0.  Not likely for
immediate float arguments to operators to ever be a possibility due to
the way instructions are stored.  It would be possible, but it would be
a lot of work, and if someone wants constants, they can just make them
static float variables.