crustris.cvm: Attempt to reduce some "high" CPU usage.
High meaning, 10% CPU on the high scores screen, which isn't high but
high for what this should be.  I cheated by using the "slow path" text
drawing routines for drawing the numbers rather than updating a reused
tilemap for outputting numbers.  I guess that's a slower operation.
This could be further improved by only updating it once per transition
to the high score table or when high scores are changed and just
displaying from that tilemap, i might consider doing that later.