paramgen.py: Give an additional useful value.
The tone instrument has 1 second long envelope buffers scaled from 0 to
1, so to apply an envelope, a start value, a difference and how "fast"
the 1 second envelope should be advanced are needed to initialize the
envelope values at the start of a note so an envelope can be applied
using the audio engine's own modulation capabilities.  The tone
instrument template just needs to be expanded to make use of these
values.