crustygame: Rewrite a lot of how the music engine works.
Not done, but at least show that there's some work being slowly done and
I guess this can act as a bit of a backup?
Anyway, try to make things considerably more dynamic and rely a lot less
on static code generation.  It was making things way too inflexible and
would've resulted in absurd amounts of code generated if there was a lot
of music, and it was just getting too janky.  Also, break up instruments
in to separate building block components.  Maybe somewhat less efficient
but it makes for far more flexibility and it shouldn't be much cost.
Incomplete and broken (although crustygame itself should still work
fine, just crustris won't run.)