instruments.inc: Remove output position tracking.
Rely on the players to keep track of the output position and just reset
them on the start of frames.  The ADSR generator doesn't need it, so it
should always just reset when called, because it's always going to be a
control for something else, and also never directly mixed.