crustris.cvm: Remove broken audio stuff.
I don't plan to continue working on crustris or adding audio to it.  I'm
most likely going to take the graphics and audio components from this
and work them in to some kind of python bindings or something because
crustyvm is just a real pain to use.  Otherwise, this is more or less in
maintenance mode.  If any bugs come up, I may fix them but I won't
likely add any new features to this project aside from maybe backporting
synthesizer or graphics component features, but for now it's basically
usable as-is.