Brought tilemap.c up to current uncrustygame tilemap.c.
Should fix some odd little bugs and things, and also just make the code
more maintainable.  It ends up being a version somewhat between old and
new to avoid breaking compatibility except in 1 case with
gfx_add_tilemap requiring a tileset be assigned to it, now.  If there's
any demand, I might consider fully upgrading it to use the same
tilemap.c as uncrustygame.

I won't bother with synth.c.  They're so disparate at this point and the
only crustygame game that exists has no audio so any bugs or problems
won't apply, but there definitely are a lot of bugs in the old synth.c.
If there's any demand, again, I might consider bringing them in-line.
Possibly even making uncrustygame a submodule of this to keep them in
sync.